Change from src/corto to cortopy and module structure
