--- src/corto/__init__.py
@@ -1,10 +1,0 @@
-from corto.rendering import camera
-from corto.rendering import body
-from corto.rendering import sun
-from corto.rendering import rendering
-from corto.rendering import state
-from corto.rendering import environment
-from corto.rendering import compositing
-
-from corto.postProcessing import PostProcessing
-
